corrects according to the reviewer comments
