+++ app/components/Community/Community.css
@@ -67,26 +67,6 @@
   font-family: Open-Sans;
 }
 
-.Community-Tags {
-  display: flex;
-  flex-direction: row;
-  margin-top: 20px;
-}
-
-.Community-Tag {
-  font-family: Open-Sans;
-  font-style: normal;
-  font-weight: 300;
-  font-size: 16px;
-  line-height: 22px;
-  background-color: white;
-  color: #2a2a2a;
-  padding: 4px 17px;
-  border: 1px solid #ababab;
-  box-sizing: border-box;
-  border-radius: 3px;
-}
-
 .Community-Description {
   margin: 40px 0;
   font-family: Open-Sans;
@@ -99,26 +79,6 @@
   height: auto;
   margin: 0 100px;
   background: #ababab;
-}
-
-.Community-Social {
-  margin-bottom: 56px;
-}
-
-.Community-SocialTitle {
-  font-family: Open-Sans;
-  font-size: 16px;
-  line-height: 24px;
-  margin-bottom: 6px;
-  color: #2a2a2a;
-}
-
-.Community-SocialIcon {
-  width: 45px;
-  height: 45px;
-  transform: scale(1.5);
-  filter: invert(57%);
-  padding: 7px;
 }
 
 .Community-Participants {
